move e2e into own folder
