+++ src/applications/gi/tests/e2e/00-required.e2e.spec.js
@@ -1,7 +1,7 @@
-const E2eHelpers = require('../../../platform/testing/e2e/helpers');
-const Timeouts = require('../../../platform/testing/e2e/timeouts');
+const E2eHelpers = require('../../../../platform/testing/e2e/helpers');
+const Timeouts = require('../../../../platform/testing/e2e/timeouts');
 const GiHelpers = require('./gibct-helpers');
-const institutionProfile = require('./data/institution-profile.json');
+const institutionProfile = require('../data/institution-profile.json');
 
 module.exports = E2eHelpers.createE2eTest(client => {
   const institutionAttributes = institutionProfile.data.attributes;
