+++ src/applications/gi/tests/e2e/01-keyboard.e2e.spec.js
@@ -1,5 +1,5 @@
-const E2eHelpers = require('../../../platform/testing/e2e/helpers');
-const Timeouts = require('../../../platform/testing/e2e/timeouts');
+const E2eHelpers = require('../../../../platform/testing/e2e/helpers');
+const Timeouts = require('../../../../platform/testing/e2e/timeouts');
 const GiHelpers = require('./gibct-helpers');
 
 module.exports = E2eHelpers.createE2eTest(client => {
