+++ src/applications/gi/tests/e2e/03-vet-tec.e2e.spec.js
@@ -1,8 +1,8 @@
-const E2eHelpers = require('../../../platform/testing/e2e/helpers');
-const Timeouts = require('../../../platform/testing/e2e/timeouts');
+const E2eHelpers = require('../../../../platform/testing/e2e/helpers');
+const Timeouts = require('../../../../platform/testing/e2e/timeouts');
 const GiHelpers = require('./gibct-helpers');
 const VetTecHelpers = require('./vet-tec-helpers');
-const vetTecSearchResults = require('./data/vet-tec-search-results.json');
+const vetTecSearchResults = require('../data/vet-tec-search-results.json');
 
 module.exports = E2eHelpers.createE2eTest(client => {
   const vetTecAttributes = vetTecSearchResults.data[0].attributes;
