+++ src/applications/gi/tests/e2e/vet-tec-helpers.js
@@ -1,9 +1,9 @@
-const Timeouts = require('../../../platform/testing/e2e/timeouts');
-const vetTecProfile = require('./data/vet-tec-profile.json');
-const vetTecSearchResults = require('./data/vet-tec-search-results.json');
-const calculatorConstants = require('./data/calculator-constants.json');
-const featureToggles = require('./data/feature-toggles.json');
-const mock = require('../../../platform/testing/e2e/mock-helpers');
+const Timeouts = require('../../../../platform/testing/e2e/timeouts');
+const vetTecProfile = require('../data/vet-tec-profile.json');
+const vetTecSearchResults = require('../data/vet-tec-search-results.json');
+const calculatorConstants = require('../data/calculator-constants.json');
+const featureToggles = require('../data/feature-toggles.json');
+const mock = require('../../../../platform/testing/e2e/mock-helpers');
 
 const vetTecAttributes = vetTecProfile.data.attributes;
